bnet/requirements.txt
torch
numpy
PyYAML
tqdm
scikit-learn
six
tensorboard